Player can rotate in circles

+++ src/graphics/include/Graphics/Drawable_I.hpp
@@ -8,9 +8,9 @@ namespace graphics
 
 struct Position_T
 {
-    int x;
-    int y;
-    int angle;
+    float x;
+    float y;
+    float angle;
 };
 
 struct Drawable_T
